Added a named list of Pantone colors and timing
Now one can choose which list of named colors should be used, but with longer lists, the time to score all listed colors increases (from ca. 5 s to 25 s for 20 colors in the pattern and 2309 in the pantone list compared to a webcolor list of 138 colors).
The time module is used to measure the elapsed time.